pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "psychotropic"
description = "A Discord bot built for harm reduction and chemistry."
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-or-later"
version = "0.0.1"
keywords = []
authors = [
  { name = "xyzt_", email = "xyzt@psychonaut.fr" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "discord.py~=2.0",
  "httpx~=0.23",
  "md-anchors~=0.1",
  "Pillow~=9.2",
]

[project.urls]
Documentation = "https://github.com/x-yzt/psychotropic#readme"
Issues = "https://github.com/x-yzt/psychotropic/issues"
Source = "https://github.com/x-yzt/psychotropic"

[tool.isort]
multi_line_output=4
lines_after_imports=2
